\name{Energy}
\alias{Energy}
\docType{data}
\title{Energy intake}
\description{
Energy intake, measured in megajoules, over a 24 hour
period of 15 year-old males and females obtained in a study on metabolism. 
}
\usage{data(Energy)}
\format{
  A data frame with 16 observations on the following 2 variables.
  \describe{
    \item{\code{Girls}}{Energy intake of 16 girls}
    \item{\code{Boys}}{Energy intake of 16 boys}
  }
}
\source{
McPherson, G.  (1990).  Statistics in Scientific Investigation.  New York:  Springer-Verlag

Chapter 7 of \strong{EnteR Monte Carlo Statistical Methods} by
Christian P. Robert and George Casella
}
\examples{
data(Energy)
str(Energy)
}
\keyword{datasets}
